Revert "Fix issue #6754"
This reverts commit ec77655b5380803c02438acc4d1be7c16c513efc.

+++ core/deployment/src/main/java/io/quarkus/deployment/pkg/steps/JarResultBuildStep.java
@@ -657,7 +657,9 @@ public class JarResultBuildStep {
                                 services.computeIfAbsent(relativePath, (u) -> new ArrayList<>()).add(content);
                             } else {
                                 Path target = fs.getPath(relativePath);
-                                Files.copy(path, target, StandardCopyOption.REPLACE_EXISTING);
+                                if (!Files.exists(target)) {
+                                    Files.copy(path, target, StandardCopyOption.REPLACE_EXISTING);
+                                }
                             }
                         }
                     } catch (Exception e) {
